backend/requirements.txt
ffmpeg-python==0.2.0
Pillow==10.4.0
